requirement.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
customtkinter==5.2.2
Cython==3.0.10
darkdetect==0.8.0
decorator==5.1.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
future==1.0.0
geocoder==1.38.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
pillow==10.3.0
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyperclip==1.8.2
ratelim==0.1.6
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.30
tk==0.1.0
tkintermapview==1.29
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
